pyproject.toml
[build-system]
requires = ["setuptools>=45", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "barcode_validator"
dynamic = ["version"]
authors = [
    {name = "Rutger Vos", email = "rutger.vos@naturalis.nl"},
]
description = "A python package for structural and taxonomic validation of DNA barcode data."
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "nbitk",
    "biopython",
    "requests",
    "PyYAML",
]

[tool.setuptools_scm]
write_to = "barcode_validator/_version.py"

[tool.setuptools.packages.find]
where = ["."]
include = ["barcode_validator*"]

[tool.setuptools.package-data]
"barcode_validator" = ["hmm_files/*.hmm"]